requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-js-asset==2.0.0
django-mptt==0.14.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==1.0.0
sqlparse==0.4.3